solis
Module core

Structs

PacketCarDamageData

This packet details car damage parameters for all the cars in the race.

PacketCarSetupData

This packet details the car setups for each vehicle in the session.

PacketCarStatusData

PacketCarTelemetryData

Telemetry for all the cars in the race.

PacketEventData

This packet gives details of events that happen during the course of a session.

PacketFinalClassificationData

This packet details the final classification at the end of the race, and the data will match with the post race results screen.

PacketHeader

Every packet will have the following header.

PacketLapData

The lap data packet gives details of all the cars in the session.

PacketLobbyInfoData

This packet details the players currently in a multiplayer lobby. It details each player’s selected car, any AI involved in the game and also the ready status of each of the participants.

PacketMotionData

Physics data for all the cars being driven.

PacketParticipantsData

This is a list of participants in the race.

PacketSessionData

PacketSessionHistoryData

This packet contains lap times and tyre usage for the session.

Enums

TelemetryPacket

A generic enum representing all possible F1 telemetry packet types

Functions

parse_packet

Attempt to create a TelemetryPacket from a buffer