solis::core
Struct PacketHeader

#[repr(C, packed(1))]
pub struct PacketHeader {
    pub packet_format: u16,
    pub game_major_version: u8,
    pub game_minor_version: u8,
    pub packet_version: u8,
    pub packet_id: u8,
    pub session_uid: u64,
    pub session_time: f32,
    pub frame_identifier: u32,
    pub player_car_index: u8,
    pub secondary_player_car_index: u8,
}

Every packet will have the following header.

Fields
packet_format: u16game_major_version: u8game_minor_version: u8packet_version: u8packet_id: u8

Identifier for the packet type, see below

Packet Name	Val	Description

Motion	0	Contains all motion data for playerโs car โ only sent while player is in control

Session	1	Data about the session โ track, time left

Lap Data	2	Data about all the lap times of cars in the session

Event	3	Various notable events that happen during a session

Participants	4	List of participants in the session, mostly relevant for multiplayer

Car Setups	5	Packet detailing car setups for cars in the race

Car Telemetry	6	Telemetry data for all cars

Car Status	7	Status data for all cars

Final Classification	8	Final classification confirmation at the end of a race

Lobby Info	9	Information about players in a multiplayer lobby

Car Damage	10	Damage status for all cars

Session History	11	Lap and tyre data for session

session_uid: u64session_time: f32frame_identifier: u32player_car_index: u8secondary_player_car_index: u8

Index of secondary playerโs car in the array (splitscreen) 255 if no second player